test(core): migrate Bucket A context/sim/util tests to commonTest (Task 3e)
Move 9 test files from jvmTest to commonTest so they run on both JVM
and linuxX64 native targets.  Extend test infrastructure as needed:

Migration (jvmTest → commonTest):
- context/BaseContextPropertyChangeTest  (11 tests, flatten 4 nested)
- context/BresenhamJoinTest              (9 tests)
- context/GridOperationsTest             (20+ tests, flatten nested)
- context/GridTransformerTest            (14 tests, flatten 6 nested)
- context/PropertyChangeTest             (10 tests)
- context/SimulationProcessFactoryTest   (3 tests)
- context/navigation/TopologyNavigatorTest (19 tests)
- sim/InOutWorkerPathHandlingTest        (14 tests, flatten 4 nested)
- util/DynamicWrapperUtilsTest           (7 tests, flatten 1 nested)

Infrastructure fixes:
- Replace org.koin.test.inject with org.koin.core.component.inject
- Replace kotlin.system.measureTimeMillis (deprecated on native)
  with kotlin.time.measureTime
- Add EditingContextFactory + TestContextBuilder bindings to
  commonCoreTestModule so commonTest DI works without jvmMain
- Add afterKoinSetUp() hook to CommonKoinTestBase; subclass @BeforeTest
  methods converted to override afterKoinSetUp() to guarantee Koin is
  started before test fixtures are constructed (Kotlin/Native ordering)
- Add containsAnyOf / hasSizeGreaterThanOrEqualTo / containsElement
  extensions to CommonAssertKExtensions

Results: JVM 1864/1864, linuxX64 671/671 (was 557 before this PR branch)
