+++ core-test/src/commonMain/kotlin/cz/vutbr/fit/interlockSim/testutil/CommonCoreTestModule.kt
@@ -2,6 +2,8 @@ package cz.vutbr.fit.interlockSim.testutil
 
 import cz.vutbr.fit.interlockSim.context.DefaultEditingContext
 import cz.vutbr.fit.interlockSim.context.DefaultSimulationContext
+import cz.vutbr.fit.interlockSim.context.EditingContext
+import cz.vutbr.fit.interlockSim.context.EditingContextFactory
 import cz.vutbr.fit.interlockSim.context.SimulationProcessFactory
 import cz.vutbr.fit.interlockSim.context.navigation.DefaultPathReservationService
 import cz.vutbr.fit.interlockSim.context.navigation.DefaultTopologyNavigator
@@ -27,9 +29,22 @@ import org.koin.dsl.module
  *
  * @since 2026-03-20 (KMP Step 4 — split commonTest for linuxX64)
  */
+/**
+ * Minimal EditingContextFactory for commonTest: only supports createEmptyContext().
+ */
+private class CommonTestEditingContextFactory : EditingContextFactory {
+	override fun createEmptyContext(): EditingContext = DefaultEditingContext(30, 30)
+}
+
 val commonCoreTestModule: Module =
 	module {
 		single<SimulationProcessFactory> { DefaultSimulationProcessFactory() }
+
+		// Provide EditingContextFactory (common-only, no XML/file I/O)
+		single<EditingContextFactory> { CommonTestEditingContextFactory() }
+
+		// Provide TestContextBuilder for tests that use buildConnectedInOut / buildMinimalSimulation
+		factory { TestContextBuilder() }
 
 		// Default editing context factory: creates a minimal linear-track context
 		factory<DefaultEditingContext> {
