--- core/src/jvmTest/kotlin/cz/vutbr/fit/interlockSim/context/GridOperationsTest.kt
@@ -1,534 +1,0 @@
-/* Brno University of Technology
- * Faculty of Information Technology
- *
- * BSc Thesis  2006/2007
- *
- * Railway Interlocking Simulator - Test Suite
- *
- * Test suite for grid operations (Issue: Context Package Test Coverage)
- */
-package cz.vutbr.fit.interlockSim.context
-
-import assertk.assertThat
-import assertk.assertions.hasSize
-import assertk.assertions.isEqualTo
-import assertk.assertions.isGreaterThan
-import assertk.assertions.isInstanceOf
-import assertk.assertions.isNotNull
-import assertk.assertions.isNull
-import assertk.assertions.isSameInstanceAs
-import cz.vutbr.fit.interlockSim.objects.cells.InOut
-import cz.vutbr.fit.interlockSim.objects.cells.RailSemaphore
-import cz.vutbr.fit.interlockSim.objects.cells.RailSwitch
-import cz.vutbr.fit.interlockSim.objects.core.Cell.SpatialType
-import cz.vutbr.fit.interlockSim.objects.tracks.SimpleTrackBlock
-import cz.vutbr.fit.interlockSim.testutil.KoinTestBase
-import cz.vutbr.fit.interlockSim.util.Point
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Nested
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
-import org.koin.test.inject
-import kotlin.test.assertFailsWith
-
-/**
- * Comprehensive tests for grid operations in DefaultEditingContext
- *
- * Test Coverage:
- * - putCell() edge cases (replacements, boundary cells, InOut tracking)
- * - removeCell() graph cleanup and InOut list updates
- * - moveCell() with various scenarios (occupied destination, invalid source)
- * - joinCells() edge cases (invalid track length, conflicting segments)
- * - Grid boundary validation
- * - Graph consistency after operations
- *
- * Total tests: 20+
- */
-@DisplayName("Grid Operations")
-class GridOperationsTest : KoinTestBase() {
-	private val editingContextFactory: EditingContextFactory by inject()
-
-	// Test cells
-	private val inA: InOut = InOut("A", false, SpatialType.HORIZONTAL)
-	private val outB: InOut = InOut("B", true, SpatialType.HORIZONTAL)
-	private val inC: InOut = InOut("C", false, SpatialType.VERTICAL)
-	private val semaphore1: RailSemaphore = RailSemaphore(false, SpatialType.HORIZONTAL)
-	private val semaphore2: RailSemaphore = RailSemaphore(true, SpatialType.DIAGONAL1)
-	private val railSwitch: RailSwitch = RailSwitch(SpatialType.HORIZONTAL, RailSwitch.Type.SIMPLE_RIGHT_FALSE)
-
-	@Nested
-	@DisplayName("putCell Operations")
-	inner class PutCellOperations {
-		@Test
-		@DisplayName("putCell replaces existing cell correctly")
-		fun putCell_replacesExistingCell() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			val point = Point(5, 5)
-
-			// Act - Add first cell
-			context.putCell(point, inA)
-			val firstCell = context.getRailWayNetGrid().getCellAt(point.x, point.y)
-
-			// Act - Replace with second cell
-			context.putCell(point, semaphore1)
-			val secondCell = context.getRailWayNetGrid().getCellAt(point.x, point.y)
-
-			// Assert - Cell was replaced
-			assertThat(firstCell).isNotNull().isInstanceOf(InOut::class)
-			assertThat(secondCell).isNotNull().isInstanceOf(RailSemaphore::class)
-			assertThat(secondCell).isSameInstanceAs(semaphore1)
-		}
-
-		@Test
-		@DisplayName("putCell at boundary (0,0) succeeds")
-		fun putCell_atBoundaryZero_succeeds() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			val point = Point(0, 0)
-
-			// Act
-			context.putCell(point, inA)
-
-			// Assert
-			val cell = context.getRailWayNetGrid().getCellAt(point.x, point.y)
-			assertThat(cell).isNotNull().isInstanceOf(InOut::class)
-		}
-
-		@Test
-		@DisplayName("putCell at boundary (max-1, max-1) succeeds")
-		fun putCell_atBoundaryMax_succeeds() {
-			// Arrange
-			val context = DefaultEditingContext(20, 20)
-			val point = Point(19, 19)
-
-			// Act
-			context.putCell(point, outB)
-
-			// Assert
-			val cell = context.getRailWayNetGrid().getCellAt(point.x, point.y)
-			assertThat(cell).isNotNull().isInstanceOf(InOut::class)
-		}
-
-		@Test
-		@DisplayName("putCell outside grid bounds throws exception")
-		fun putCell_outsideBounds_throwsException() {
-			// Arrange
-			val context = DefaultEditingContext(20, 20)
-
-			// Act & Assert - X too large
-			assertThrows<ContextCreationException> {
-				context.putCell(Point(20, 5), inA)
-			}
-
-			// Act & Assert - Y too large
-			assertThrows<ContextCreationException> {
-				context.putCell(Point(5, 20), outB)
-			}
-
-			// Act & Assert - Negative X
-			assertThrows<ContextCreationException> {
-				context.putCell(Point(-1, 5), inC)
-			}
-
-			// Act & Assert - Negative Y
-			assertThrows<ContextCreationException> {
-				context.putCell(Point(5, -1), semaphore1)
-			}
-		}
-
-		@Test
-		@DisplayName("putCell adds InOut to InOut list")
-		fun putCell_addsInOutToList() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-
-			// Act
-			context.putCell(Point(1, 1), inA)
-			context.putCell(Point(5, 5), outB)
-			context.putCell(Point(3, 3), semaphore1) // Not an InOut
-
-			// Assert - Only InOuts are in the list
-			val inOuts = context.getInOuts()
-			assertThat(inOuts).hasSize(2)
-		}
-
-		@Test
-		@DisplayName("putCell creates automatic track blocks for adjacent cells")
-		fun putCell_createsAutomaticTrackBlocks() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-
-			// Act - Place two horizontally adjacent InOuts
-			context.putCell(Point(5, 5), inA)
-			context.putCell(Point(6, 5), outB)
-
-			// Assert - Graph should contain automatic connection
-			val graphSize = context.getGraph().size()
-			assertThat(graphSize).isGreaterThan(0)
-		}
-	}
-
-	@Nested
-	@DisplayName("removeCell Operations")
-	inner class RemoveCellOperations {
-		@Test
-		@DisplayName("removeCell updates graph connections")
-		fun removeCell_updatesGraphConnections() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			val point = Point(5, 5)
-			context.putCell(point, inA)
-			context.putCell(Point(6, 5), outB)
-			val graphSizeBefore = context.getGraph().size()
-
-			// Act - Remove cell
-			context.removeCell(point)
-
-			// Assert - Graph entries related to removed cell are gone
-			val graphSizeAfter = context.getGraph().size()
-			val cell = context.getRailWayNetGrid().getCellAt(point.x, point.y)
-			assertThat(cell).isNull()
-			assertThat(graphSizeAfter).isEqualTo(0) // All connections to removed cell should be gone
-		}
-
-		@Test
-		@DisplayName("removeCell on empty cell does nothing")
-		fun removeCell_emptyCell_doesNothing() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			val point = Point(5, 5)
-
-			// Act - Remove from empty location (should not throw)
-			context.removeCell(point)
-
-			// Assert - No exception thrown
-			assertThat(context.getRailWayNetGrid().getCellAt(point.x, point.y)).isNull()
-		}
-
-		@Test
-		@DisplayName("removeCell removes InOut from InOut list")
-		fun removeCell_removesInOutFromList() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			context.putCell(Point(1, 1), inA)
-			context.putCell(Point(5, 5), outB)
-			assertThat(context.getInOuts()).hasSize(2)
-
-			// Act - Remove one InOut
-			context.removeCell(Point(1, 1))
-
-			// Assert - InOut list updated
-			val inOuts = context.getInOuts()
-			assertThat(inOuts).hasSize(1)
-		}
-
-		@Test
-		@DisplayName("removeCell removes intermediate track block parts")
-		fun removeCell_removesTrackBlockParts() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			context.putCell(Point(1, 1), inA)
-			context.putCell(Point(10, 10), outB)
-			val trackBlock = SimpleTrackBlock(inA, outB, 1000.0, 80.0)
-			context.joinCells(Point(1, 1), Point(10, 10), trackBlock)
-
-			// Count grid cells before removal
-			val cellCountBefore =
-				context
-					.getRailWayNetGrid()
-					.iterator()
-					.asSequence()
-					.count()
-
-			// Act - Remove one endpoint
-			context.removeCell(Point(1, 1))
-
-			// Assert - Intermediate cells should also be removed
-			val cellCountAfter =
-				context
-					.getRailWayNetGrid()
-					.iterator()
-					.asSequence()
-					.count()
-			assertThat(cellCountAfter).isEqualTo(1) // Only outB remains
-		}
-	}
-
-	@Nested
-	@DisplayName("moveCell Operations")
-	inner class MoveCellOperations {
-		@Test
-		@DisplayName("moveCell preserves track connections")
-		fun moveCell_preservesTrackConnections() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			val from = Point(5, 5)
-			val to = Point(7, 7)
-			context.putCell(from, inA)
-			context.putCell(Point(6, 5), outB)
-
-			// Act
-			context.moveCell(from, to)
-
-			// Assert - Cell moved to new location
-			assertThat(context.getRailWayNetGrid().getCellAt(from.x, from.y)).isNull()
-			assertThat(context.getRailWayNetGrid().getCellAt(to.x, to.y)).isNotNull()
-			assertThat(context.getRailWayNetGrid().getCellAt(to.x, to.y)).isSameInstanceAs(inA)
-		}
-
-		@Test
-		@DisplayName("moveCell to occupied destination does nothing")
-		fun moveCell_occupiedDestination_doesNothing() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			val from = Point(5, 5)
-			val to = Point(7, 7)
-			context.putCell(from, inA)
-			context.putCell(to, outB)
-
-			// Act - Try to move to occupied cell
-			context.moveCell(from, to)
-
-			// Assert - Source cell still at original location
-			assertThat(context.getRailWayNetGrid().getCellAt(from.x, from.y)).isSameInstanceAs(inA)
-			assertThat(context.getRailWayNetGrid().getCellAt(to.x, to.y)).isSameInstanceAs(outB)
-		}
-
-		@Test
-		@DisplayName("moveCell from empty source does nothing")
-		fun moveCell_emptySource_doesNothing() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			val from = Point(5, 5)
-			val to = Point(7, 7)
-
-			// Act - Try to move from empty cell
-			context.moveCell(from, to)
-
-			// Assert - Nothing happens
-			assertThat(context.getRailWayNetGrid().getCellAt(from.x, from.y)).isNull()
-			assertThat(context.getRailWayNetGrid().getCellAt(to.x, to.y)).isNull()
-		}
-
-		@Test
-		@DisplayName("moveCell maintains InOut list")
-		fun moveCell_maintainsInOutList() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			context.putCell(Point(5, 5), inA)
-			assertThat(context.getInOuts()).hasSize(1)
-
-			// Act - Move InOut
-			context.moveCell(Point(5, 5), Point(7, 7))
-
-			// Assert - InOut list unchanged
-			assertThat(context.getInOuts()).hasSize(1)
-		}
-	}
-
-	@Nested
-	@DisplayName("joinCells Operations")
-	inner class JoinCellsOperations {
-		@Test
-		@DisplayName("joinCells with invalid track length fails gracefully")
-		fun joinCells_invalidTrackLength_failsGracefully() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			context.putCell(Point(1, 1), inA)
-			context.putCell(Point(10, 10), outB)
-
-			// Act & Assert - Creating track block with invalid length throws IllegalArgumentException
-			assertFailsWith<IllegalArgumentException> {
-				SimpleTrackBlock(inA, outB, -100.0, 80.0)
-			}
-		}
-
-		@Test
-		@DisplayName("joinCells adjacent cells succeeds")
-		fun joinCells_adjacentCells_succeeds() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			context.putCell(Point(5, 5), inA)
-			context.putCell(Point(6, 5), outB)
-
-			// Act - Join adjacent cells
-			val trackBlock = SimpleTrackBlock(inA, outB, 100.0, 80.0)
-			context.joinCells(Point(5, 5), Point(6, 5), trackBlock)
-
-			// Assert - Graph contains the connection
-			assertThat(context.getGraph().size()).isGreaterThan(0)
-		}
-
-		@Test
-		@DisplayName("joinCells distant cells creates intermediate parts")
-		fun joinCells_distantCells_createsIntermediateParts() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			context.putCell(Point(1, 1), inA)
-			context.putCell(Point(10, 10), outB)
-
-			// Count cells before join
-			val cellCountBefore =
-				context
-					.getRailWayNetGrid()
-					.iterator()
-					.asSequence()
-					.count()
-
-			// Act - Join distant cells
-			val trackBlock = SimpleTrackBlock(inA, outB, 1000.0, 80.0)
-			context.joinCells(Point(1, 1), Point(10, 10), trackBlock)
-
-			// Count cells after join
-			val cellCountAfter =
-				context
-					.getRailWayNetGrid()
-					.iterator()
-					.asSequence()
-					.count()
-
-			// Assert - Intermediate cells created (if join succeeded)
-			// Note: Join might fail if cells are too distant or incompatible
-			// This test verifies the attempt doesn't crash
-		}
-
-		@Test
-		@DisplayName("joinCells with conflicting segments fails")
-		fun joinCells_conflictingSegments_fails() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			// Place cells with incompatible orientations
-			val verticalInOut = InOut("V", false, SpatialType.VERTICAL)
-			val horizontalInOut = InOut("H", false, SpatialType.HORIZONTAL)
-			context.putCell(Point(5, 5), verticalInOut)
-			context.putCell(Point(6, 5), horizontalInOut)
-
-			// Act - Try to join incompatible cells
-			val trackBlock = SimpleTrackBlock(verticalInOut, horizontalInOut, 100.0, 80.0)
-			context.joinCells(Point(5, 5), Point(6, 5), trackBlock)
-
-			// Assert - Join should fail (implementation might emit JOIN_FAILED event)
-		}
-	}
-
-	@Nested
-	@DisplayName("removeLine Operations")
-	inner class RemoveLineOperations {
-		@Test
-		@DisplayName("removeLine removes track block and intermediate cells")
-		fun removeLine_removesTrackBlockAndIntermediateCells() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			context.putCell(Point(1, 1), inA)
-			context.putCell(Point(10, 10), outB)
-			val trackBlock = SimpleTrackBlock(inA, outB, 1000.0, 80.0)
-			context.joinCells(Point(1, 1), Point(10, 10), trackBlock)
-
-			val cellCountBefore =
-				context
-					.getRailWayNetGrid()
-					.iterator()
-					.asSequence()
-					.count()
-			val graphSizeBefore = context.getGraph().size()
-
-			// Act
-			context.removeLine(trackBlock)
-
-			// Assert
-			val cellCountAfter =
-				context
-					.getRailWayNetGrid()
-					.iterator()
-					.asSequence()
-					.count()
-			val graphSizeAfter = context.getGraph().size()
-
-			// Graph should have fewer entries
-			assertThat(graphSizeAfter).isEqualTo(0) // Track block removed
-
-			// Intermediate cells should be removed (only InOuts remain)
-			assertThat(cellCountAfter).isEqualTo(2) // Only inA and outB remain
-		}
-
-		@Test
-		@DisplayName("removeLine on non-existent track does nothing")
-		fun removeLine_nonExistentTrack_doesNothing() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			context.putCell(Point(1, 1), inA)
-			context.putCell(Point(5, 5), outB)
-			val trackBlock = SimpleTrackBlock(inA, outB, 100.0, 80.0)
-			// Note: We DON'T call joinCells, so trackBlock is not in the graph
-
-			// Act - Try to remove non-existent track (should not throw)
-			context.removeLine(trackBlock)
-
-			// Assert - No exception thrown
-			assertThat(context.getGraph().size()).isEqualTo(0)
-		}
-	}
-
-	@Nested
-	@DisplayName("Grid Consistency")
-	inner class GridConsistency {
-		@Test
-		@DisplayName("grid maintains location-to-cell mapping consistency")
-		fun grid_maintainsLocationCellMapping() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			val point = Point(5, 5)
-
-			// Act
-			context.putCell(point, inA)
-
-			// Assert - Cell can be retrieved by location
-			val cell = context.getRailWayNetGrid().getCellAt(point.x, point.y)
-			assertThat(cell).isSameInstanceAs(inA)
-
-			// Assert - Location can be retrieved by cell
-			val location = context.getRailWayNetGrid().getLocation(inA)
-			assertThat(location).isEqualTo(point)
-		}
-
-		@Test
-		@DisplayName("grid iterator reflects all cells")
-		fun grid_iteratorReflectsAllCells() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			context.putCell(Point(1, 1), inA)
-			context.putCell(Point(5, 5), outB)
-			context.putCell(Point(3, 3), semaphore1)
-
-			// Act
-			val cells =
-				context
-					.getRailWayNetGrid()
-					.iterator()
-					.asSequence()
-					.toList()
-
-			// Assert - All cells are in the iterator
-			assertThat(cells).hasSize(3)
-		}
-
-		@Test
-		@DisplayName("graph size matches track block count")
-		fun graph_sizeMatchesTrackBlockCount() {
-			// Arrange
-			val context = editingContextFactory.createEmptyContext()
-			context.putCell(Point(1, 1), inA)
-			context.putCell(Point(2, 1), outB)
-			context.putCell(Point(5, 5), inC)
-
-			// Act - Create two track blocks
-			val trackBlock1 = SimpleTrackBlock(inA, outB, 100.0, 80.0)
-			context.joinCells(Point(1, 1), Point(2, 1), trackBlock1)
-
-			// Assert - Graph size reflects track blocks
-			val graphSize = context.getGraph().size()
-			assertThat(graphSize).isGreaterThan(0)
-		}
-	}
-}
