--- core/src/jvmTest/kotlin/cz/vutbr/fit/interlockSim/sim/InOutWorkerPathHandlingTest.kt
@@ -1,285 +1,0 @@
-/* Brno University of Technology
- * Faculty of Information Technology
- *
- * BSc Thesis  2006/2007
- *
- * Railway Interlocking Simulator - Test Suite
- *
- * Phase 2.3: InOutWorker Path Handling Tests
- */
-package cz.vutbr.fit.interlockSim.sim
-
-import assertk.assertThat
-import assertk.assertions.isFalse
-import assertk.assertions.isNotNull
-import assertk.assertions.isTrue
-import cz.vutbr.fit.interlockSim.context.SimulationContext
-import cz.vutbr.fit.interlockSim.objects.cells.DynamicInOut
-import cz.vutbr.fit.interlockSim.objects.cells.RailSemaphore
-import cz.vutbr.fit.interlockSim.objects.core.Cell.SpatialType
-import cz.vutbr.fit.interlockSim.testutil.KoinTestBase
-import cz.vutbr.fit.interlockSim.testutil.MockSimulationContext
-import cz.vutbr.fit.interlockSim.testutil.withMessage
-import cz.hovorka.kdisco.Head
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Nested
-import org.junit.jupiter.api.Test
-
-/**
- * Unit tests for path and semaphore coordination in InOutWorker.
- *
- * Coverage (12 tests, ~200 instructions):
- * - Path setup coordination (4 tests)
- * - Semaphore coordination (3 tests)
- * - Queue management edge cases (5 tests)
- *
- * These tests validate InOutWorker's capability to:
- * - Coordinate path setup before allowing trains to enter
- * - Manage semaphore states during path operations
- * - Handle queue edge cases (empty queue, FIFO order, train removal)
- *
- * Note: Full simulation execution tests (waitUntil, actual path reservation)
- * require kDisco framework and are beyond scope of unit testing.
- */
-class InOutWorkerPathHandlingTest : KoinTestBase() {
-	private lateinit var context: MockSimulationContext
-	private lateinit var entryInOut: DynamicInOut
-	private lateinit var worker: InOutWorker
-	private lateinit var queue: Head
-
-	@BeforeEach
-	fun setUp() {
-		// Create a context with a connected InOut
-		val (ctx, inOut) = createTestContextWithInOut("TEST_ENTRY", false, SpatialType.HORIZONTAL)
-		context = MockSimulationContext(ctx as cz.vutbr.fit.interlockSim.context.DefaultSimulationContext)
-		testContext = context // Track for cleanup
-		entryInOut = inOut
-		worker = InOutWorker(context, entryInOut)
-		queue = worker.getQueqe()
-
-		// Verify initial state
-		assertThat(queue.empty())
-			.withMessage("Queue should be empty after worker construction")
-			.isTrue()
-	}
-
-	/**
-	 * Helper function to create a test context with a connected DynamicInOut.
-	 */
-	private fun createTestContextWithInOut(
-		name: String,
-		orientation: Boolean,
-		spatialType: SpatialType
-	): Pair<SimulationContext, DynamicInOut> {
-		// Create a context with a connected InOut
-		val context =
-			cz.vutbr.fit.interlockSim.testutil.buildConnectedInOut(
-				inOutName = name,
-				isEntry = !orientation // orientation is inverted: false = entry, true = exit
-			)
-
-		// Return the context and the first InOut
-		val inOut = context.getInOuts().first()
-		return Pair(context, inOut)
-	}
-
-	@Nested
-	@DisplayName("Path Setup Coordination")
-	inner class PathSetupCoordinationTests {
-		@Test
-		fun `worker sets up path before train entry`() {
-			// Worker waits for pathFree condition, then calls path.setUpPath(inOut)
-			// This ensures path is available before granting train entry
-
-			assertThat(worker)
-				.withMessage("Worker should be initialized for path coordination")
-				.isNotNull()
-		}
-
-		@Test
-		fun `worker waits for path availability`() {
-			// In iteration(), worker: waitUntil(pathFree)
-			// pathFree condition checks if path exists and is free from inOut
-			// Create a properly connected worker for this test
-
-			val (orphanContext, orphanInOut) = createTestContextWithInOut("ORPHAN", false, SpatialType.VERTICAL)
-			val orphanWorker = InOutWorker(orphanContext, orphanInOut)
-
-			assertThat(orphanWorker)
-				.withMessage("Worker should handle connected InOut")
-				.isNotNull()
-		}
-
-		@Test
-		fun `worker handles path setup failure gracefully`() {
-			// If path.setUpPath() throws exception (e.g., TrackOperationException),
-			// worker catches it, logs APPROVAL_DENIED, calls context.errorStop()
-
-			val (failureContext, failureInOut) = createTestContextWithInOut("FAILURE_TEST", false, SpatialType.HORIZONTAL)
-			val failureWorker = InOutWorker(failureContext, failureInOut)
-
-			assertThat(failureWorker)
-				.withMessage("Worker should be ready to handle path setup failures")
-				.isNotNull()
-		}
-
-		@Test
-		fun `worker releases path after train exits`() {
-			// After train leaves queue (first != queqe.first()),
-			// loop continues to next train or exits
-			// Path is implicitly released when not holding reference
-
-			val releaseTest = worker.getQueqe()
-
-			assertThat(releaseTest)
-				.withMessage("Queue should support path release for next train")
-				.isNotNull()
-		}
-	}
-
-	@Nested
-	@DisplayName("Semaphore Coordination")
-	inner class SemaphoreCoordinationTests {
-		@Test
-		fun `worker sets semaphore to proceed when path ready`() {
-			// path.setUpPath(inOut) configures entry semaphore via setUpSemaphores()
-			// Entry semaphore (getLastPathSemaphore) transitions to PROCEED aspect
-			// Allows train to enter from path
-
-			val testSemaphore = RailSemaphore(false, SpatialType.HORIZONTAL)
-
-			assertThat(testSemaphore)
-				.withMessage("Semaphore should be configured when path is set up")
-				.isNotNull()
-		}
-
-		@Test
-		fun `worker resets semaphore after train passes`() {
-			// When train leaves queue, worker moves to next train
-			// Previous path's semaphore resets when path.cancelPathSetup() called
-			// (or implicitly when not holding reference anymore)
-
-			val (entryContext, entryPoint) = createTestContextWithInOut("ENTRY", false, SpatialType.HORIZONTAL)
-			val entryWorker = InOutWorker(entryContext, entryPoint)
-
-			assertThat(entryWorker.getQueqe())
-				.withMessage("Worker should reset semaphore after train processing")
-				.isNotNull()
-		}
-
-		@Test
-		fun `worker coordinates multiple semaphores on path`() {
-			// Path contains sequence of PathElements including multiple RailSemaphores
-			// setUpSemaphores() backtracks and configures each semaphore for route
-			// Worker manages this via path operations
-
-			val (complexContext, complexEntry) = createTestContextWithInOut("COMPLEX_ENTRY", false, SpatialType.HORIZONTAL)
-			val complexWorker = InOutWorker(complexContext, complexEntry)
-
-			assertThat(complexWorker.getQueqe())
-				.withMessage("Worker should coordinate multiple semaphores")
-				.isNotNull()
-		}
-	}
-
-	@Nested
-	@DisplayName("Queue Management Edge Cases")
-	inner class QueueManagementEdgeCases {
-		@Test
-		fun `worker handles empty queue`() {
-			// iteration(): while (!queqe.empty()) {...}
-			// If queue empty, loop doesn't execute, worker becomes idle
-			// myIdle = true, waiting for trains via enterTrain() activation
-
-			assertThat(queue.empty())
-				.withMessage("Queue should be empty initially")
-				.isTrue()
-
-			val remainsEmpty = queue.empty()
-			assertThat(remainsEmpty)
-				.withMessage("Empty queue should remain empty")
-				.isTrue()
-		}
-
-		@Test
-		fun `worker handles queue overflow`() {
-			// Head (kDisco linked list) has no fixed capacity limit
-			// Trains can queue indefinitely as they enter via Process.wait(queqe)
-			// Worker processes them one by one as paths become available
-
-			assertThat(queue)
-				.withMessage("Queue should support arbitrary number of waiting trains")
-				.isNotNull()
-		}
-
-		@Test
-		fun `worker maintains FIFO order`() {
-			// Queue.first() returns first train in queue
-			// Trains are linked via kDisco Link structure
-			// FIFO order preserved: when first train leaves, next becomes head
-
-			val queue2 = worker.getQueqe()
-
-			assertThat(queue2 == queue)
-				.withMessage("Worker maintains reference to same queue instance")
-				.isTrue()
-		}
-
-		@Test
-		fun `worker handles train removed from queue`() {
-			// If train is removed (via Link.out() or emergency stop),
-			// queue continues with remaining trains
-			// Worker's waitUntil { first != queqe.first() } detects change
-
-			val recoveryQueue = worker.getQueqe()
-
-			assertThat(recoveryQueue)
-				.withMessage("Queue should support dynamic train removal and recovery")
-				.isNotNull()
-		}
-
-		@Test
-		fun `worker handles concurrent train requests`() {
-			// Multiple InOuts can have workers with separate queues
-			// Each worker processes independently
-			// Trains entering via enterTrain() are serialized per worker
-
-			val (context1, entry1) = createTestContextWithInOut("ENTRY_1", false, SpatialType.HORIZONTAL)
-			val (context2, entry2) = createTestContextWithInOut("ENTRY_2", false, SpatialType.VERTICAL)
-
-			val worker1 = InOutWorker(context1, entry1)
-			val worker2 = InOutWorker(context2, entry2)
-
-			assertThat(worker1.getQueqe() == worker2.getQueqe())
-				.withMessage("Concurrent requests to different workers use separate queues")
-				.isFalse()
-		}
-	}
-
-	@Nested
-	@DisplayName("Worker-Path-Semaphore Coordination")
-	inner class CoordinationIntegrationTests {
-		@Test
-		fun `worker queue is available for train processing`() {
-			// getQueqe() returns queue for train entry/exit monitoring
-			val processingQueue = worker.getQueqe()
-
-			assertThat(processingQueue)
-				.withMessage("Queue should be accessible for train coordination")
-				.isNotNull()
-		}
-
-		@Test
-		fun `worker enters iteration when queue non-empty`() {
-			// iteration() called by kDisco framework when worker activated
-			// Condition: while (!queqe.empty()) ensures trains are processed
-
-			val loopCondition = !queue.empty()
-
-			assertThat(loopCondition)
-				.withMessage("Loop should not execute while queue is empty")
-				.isFalse()
-		}
-	}
-}
